Add lib definitions to example 6 bin.spdx file
Fixes #117

Note that it may be desireable to express a relationship from an
external element - however in the SPDX 2.X model, relationships are
properties of an element and adding properties to an external element
would be lost or cause an error.

Note that this is not an issue for SPDX 3.0.

See related https://github.com/spdx/spdx-java-model-2_X/issues/17